+++ google2/src/main/scala/org/broadinstitute/dsde/workbench/google2/ComputePollOperationInterpreter.scala
@@ -9,23 +9,23 @@ import org.broadinstitute.dsde.workbench.model.google.GoogleProject
 import org.typelevel.log4cats.StructuredLogger
 
 class ComputePollOperationInterpreter[F[_]: StructuredLogger](
-  zoneOperationClient: ZoneOperationClient,
-  regionOperationClient: RegionOperationClient,
-  globalOperationClient: GlobalOperationClient,
+  zoneOperationClient: ZoneOperationsClient,
+  regionOperationClient: RegionOperationsClient,
+  globalOperationClient: GlobalOperationsClient,
   blockerBound: Semaphore[F]
 )(implicit override val F: Async[F])
     extends ComputePollOperation[F] {
   override def getZoneOperation(project: GoogleProject, zoneName: ZoneName, operationName: OperationName)(implicit
     ev: Ask[F, TraceId]
   ): F[Operation] = {
-    val request = ProjectZoneOperationName
+    val request = GetZoneOperationRequest
       .newBuilder()
       .setProject(project.value)
       .setZone(zoneName.value)
       .setOperation(operationName.value)
       .build
     tracedLogging(
-      blockOn( F.blocking(zoneOperationClient.getZoneOperation(request))),
+      blockOn( F.blocking(zoneOperationClient.get(request))),
       s"com.google.cloud.compute.v1.ZoneOperationClient.getZoneOperation(${request.toString})",
       showOperation
     )
@@ -34,14 +34,14 @@ class ComputePollOperationInterpreter[F[_]: StructuredLogger](
   override def getRegionOperation(project: GoogleProject, regionName: RegionName, operationName: OperationName)(implicit
     ev: Ask[F, TraceId]
   ): F[Operation] = {
-    val request = ProjectRegionOperationName
+    val request = GetRegionOperationRequest
       .newBuilder()
       .setProject(project.value)
       .setRegion(regionName.value)
       .setOperation(operationName.value)
       .build
     tracedLogging(
-      blockOn(F.blocking(regionOperationClient.getRegionOperation(request))),
+      blockOn(F.blocking(regionOperationClient.get(request))),
       s"com.google.cloud.compute.v1.regionOperationClient.getRegionOperation(${request.toString})",
       showOperation
     )
@@ -51,9 +51,9 @@ class ComputePollOperationInterpreter[F[_]: StructuredLogger](
     ev: Ask[F, TraceId]
   ): F[Operation] = {
     val request =
-      ProjectGlobalOperationName.newBuilder().setProject(project.value).setOperation(operationName.value).build
+      GetGlobalOperationRequest.newBuilder().setProject(project.value).setOperation(operationName.value).build
     tracedLogging(
-      blockOn(F.blocking(globalOperationClient.getGlobalOperation(request))),
+      blockOn(F.blocking(globalOperationClient.get(request))),
       s"com.google.cloud.compute.v1.globalOperationClient.getGlobalOperation(${request.toString})",
       showOperation
     )