+++ google2/src/test/scala/org/broadinstitute/dsde/workbench/google2/mock/FakeGoogleComputeService.scala
@@ -23,12 +23,12 @@ class FakeGoogleComputeService extends GoogleComputeService[IO] {
   override def createInstance(project: GoogleProject, zone: ZoneName, instance: Instance)(implicit
     ev: Ask[IO, TraceId]
   ): IO[Option[Operation]] =
-    IO.pure(Some(Operation.newBuilder().setId("op").setName("opName").setTargetId("target").build()))
+    IO.pure(Some(Operation.newBuilder().setId(123).setName("opName").setTargetId(258165385).build()))
 
   override def deleteInstance(project: GoogleProject, zone: ZoneName, instanceName: InstanceName)(implicit
     ev: Ask[IO, TraceId]
   ): IO[Option[Operation]] =
-    IO.pure(Some(Operation.newBuilder().setId("op").setName("opName").setTargetId("target").build()))
+    IO.pure(Some(Operation.newBuilder().setId(123).setName("opName").setTargetId(258165385).build()))
 
   override def getInstance(project: GoogleProject, zone: ZoneName, instanceName: InstanceName)(implicit
     ev: Ask[IO, TraceId]
@@ -36,11 +36,11 @@ class FakeGoogleComputeService extends GoogleComputeService[IO] {
 
   override def stopInstance(project: GoogleProject, zone: ZoneName, instanceName: InstanceName)(implicit
     ev: Ask[IO, TraceId]
-  ): IO[Operation] = IO.pure(Operation.newBuilder().setId("op").setName("opName").setTargetId("target").build())
+  ): IO[Operation] = IO.pure(Operation.newBuilder().setId(123).setName("opName").setTargetId(258165385).build())
 
   override def startInstance(project: GoogleProject, zone: ZoneName, instanceName: InstanceName)(implicit
     ev: Ask[IO, TraceId]
-  ): IO[Operation] = IO.pure(Operation.newBuilder().setId("op").setName("opName").setTargetId("target").build())
+  ): IO[Operation] = IO.pure(Operation.newBuilder().setId(123).setName("opName").setTargetId(258165385).build())
 
   override def modifyInstanceMetadata(
     project: GoogleProject,
@@ -53,7 +53,7 @@ class FakeGoogleComputeService extends GoogleComputeService[IO] {
 
   override def addFirewallRule(project: GoogleProject, firewall: Firewall)(implicit
     ev: Ask[IO, TraceId]
-  ): IO[Operation] = IO.pure(Operation.newBuilder().setId("op").setName("opName").setTargetId("target").build())
+  ): IO[Operation] = IO.pure(Operation.newBuilder().setId(123).setName("opName").setTargetId(258165385).build())
 
   override def getFirewallRule(project: GoogleProject, firewallRuleName: FirewallRuleName)(implicit
     ev: Ask[IO, TraceId]
@@ -84,7 +84,7 @@ class FakeGoogleComputeService extends GoogleComputeService[IO] {
 
   override def createNetwork(project: GoogleProject, network: Network)(implicit
     ev: Ask[IO, TraceId]
-  ): IO[Operation] = IO.pure(Operation.newBuilder().setId("op").setName("opName").setTargetId("target").build())
+  ): IO[Operation] = IO.pure(Operation.newBuilder().setId(123).setName("opName").setTargetId(258165385).build())
 
   override def getSubnetwork(project: GoogleProject, region: RegionName, subnetwork: SubnetworkName)(implicit
     ev: Ask[IO, TraceId]
@@ -92,12 +92,12 @@ class FakeGoogleComputeService extends GoogleComputeService[IO] {
 
   override def createSubnetwork(project: GoogleProject, region: RegionName, subnetwork: Subnetwork)(implicit
     ev: Ask[IO, TraceId]
-  ): IO[Operation] = IO.pure(Operation.newBuilder().setId("op").setName("opName").setTargetId("target").build())
+  ): IO[Operation] = IO.pure(Operation.newBuilder().setId(123).setName("opName").setTargetId(258165385).build())
 
   def detachDisk(project: GoogleProject, zone: ZoneName, instanceName: InstanceName, deviceName: DeviceName)(implicit
     ev: Ask[IO, TraceId]
   ): IO[Option[Operation]] =
-    IO.pure(Some(Operation.newBuilder().setId("op").setName("opName").setTargetId("target").build()))
+    IO.pure(Some(Operation.newBuilder().setId(123).setName("opName").setTargetId(258165385).build()))
 
   override def deleteInstanceWithAutoDeleteDisk(
     project: GoogleProject,
@@ -105,7 +105,7 @@ class FakeGoogleComputeService extends GoogleComputeService[IO] {
     instanceName: InstanceName,
     autoDeleteDisks: Set[DiskName]
   )(implicit ev: Ask[IO, TraceId], computePollOperation: ComputePollOperation[IO]): IO[Option[Operation]] =
-    IO.pure(Some(Operation.newBuilder().setId("op").setName("opName").setStatus("DONE").setTargetId("target").build()))
+    IO.pure(Some(Operation.newBuilder().setId(123).setName("opName").setStatus(Operation.Status.DONE).setTargetId(258165385).build()))
 }
 
 object FakeGoogleComputeService extends FakeGoogleComputeService